Refactor: 코드 의존성 정리
