+++ src/main/java/com/project/doubleshop/domain/address/entity/Address.java
@@ -13,7 +13,7 @@ import javax.persistence.ManyToOne;
 
 import com.project.doubleshop.domain.common.Status;
 import com.project.doubleshop.domain.common.StatusConverter;
-import com.project.doubleshop.domain.member.entity.Member;
+import com.project.doubleshop.domain.member.domain.Member;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
