+++ src/main/java/com/project/doubleshop/domain/cart/service/CartService.java
@@ -16,8 +16,8 @@ import com.project.doubleshop.domain.exception.BadRequestException;
 import com.project.doubleshop.domain.exception.NotFoundException;
 import com.project.doubleshop.domain.item.entity.Item;
 import com.project.doubleshop.domain.item.service.ItemService;
-import com.project.doubleshop.domain.member.entity.Member;
-import com.project.doubleshop.domain.member.service.MemberService;
+import com.project.doubleshop.domain.member.domain.Member;
+import com.project.doubleshop.domain.member.application.MemberService;
 import com.project.doubleshop.domain.utils.ExceptionUtils;
 import com.project.doubleshop.web.cart.dto.CartDto;
 import com.project.doubleshop.web.cart.dto.CartItem;
