+++ src/main/java/com/project/doubleshop/domain/member/infrastructure/token/SimpleAuthenticationToken.java
@@ -5,7 +5,7 @@ import java.util.Collection;
 import org.springframework.security.authentication.AbstractAuthenticationToken;
 import org.springframework.security.core.GrantedAuthority;
 
-import com.project.doubleshop.web.member.dto.AuthenticationRequest;
+import com.project.doubleshop.domain.member.presentation.request.AuthenticationRequest;
 
 public class SimpleAuthenticationToken extends AbstractAuthenticationToken {
