+++ src/main/java/com/project/doubleshop/web/member/dto/ResultRole.java
@@ -1,6 +1,6 @@
 package com.project.doubleshop.web.member.dto;
 
-import com.project.doubleshop.web.config.security.Role;
+import com.project.doubleshop.domain.member.infrastructure.token.Role;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
