+++ src/main/java/com/project/doubleshop/web/member/dto/UpdateMemberRequest.java
@@ -1,10 +1,5 @@
 package com.project.doubleshop.web.member.dto;
 
-import java.time.LocalDateTime;
-
-import com.project.doubleshop.domain.common.Status;
-import com.project.doubleshop.domain.member.entity.Member;
-
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
