+++ src/main/resources/mapper/authenticationMemberMapper.xml
@@ -1,7 +1,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.project.doubleshop.domain.member.mapper.AuthMemberMapper">
+<mapper namespace="com.project.doubleshop.domain.member.infrastructure.myBatis.mapper.AuthMemberMapper">
     <insert id="insertMember" parameterType="Member" useGeneratedKeys="true" keyProperty="id">
         INSERT INTO MEMBER (user_id, password, name, email,
                             phone, count, last_login_time, status,
